Final commit

+++ routes/books.js
@@ -1,6 +1,21 @@
 var express = require('express');
 const axios = require('axios');
+var jwt = require('jsonwebtoken');
 var booksRouter = express.Router();
+
+const authUser = (req, res, next) => {
+  const authHeader = req.headers['authorization'];
+  const token = authHeader && authHeader.split(' ')[1];
+  if(!token) 
+    return res.status(401).send("Unauthenticated.");
+
+  jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+    if(err) 
+      return res.status(403).send("Expired Token.");
+    req.user = user;
+    next();
+  })
+};
 
 // Task 1: Get the book list available in the shop
 booksRouter.get('/', async (req, res, next) => {
@@ -49,40 +64,57 @@ booksRouter.get('/reviews/:reviewId', async (req, res, next) => {
 
 // Registered Users
 // Task 8: Add/Modify a book review
-booksRouter.post('/:bookId/reviews', async (req, res, next) => {
-  res.setHeader('Content-Type', 'text/plain');
-  await axios.post(
-    'http://localhost:3001/reviews', 
-    {
-      username: "empty",
-      bookId: req.params.bookId,
-      review: req.body.review
-    }
-  );
-  res.status(200).send("Review Added.");
-});
+booksRouter.post(
+  '/:bookId/reviews', 
+  authUser,
+  async (req, res, next) => {
+    res.setHeader('Content-Type', 'text/plain');
+    await axios.post(
+      'http://localhost:3001/reviews', 
+      {
+        username: req.user.name,
+        bookId: req.params.bookId,
+        review: req.body.review
+      }
+    );
+    res.status(200).send("Review Added.");
+  }
+);
 
-booksRouter.put('/:bookId/reviews/:reviewId', async (req, res, next) => {
-  res.setHeader('Content-Type', 'text/plain');
-  await axios.put(
-    'http://localhost:3001/reviews/7', 
-    {
-      username: "empty",
-      bookId: req.params.bookId,
-      review: req.body.review
-    }
-  );
-  res.status(200).send("Review Added.");
-});
+booksRouter.put(
+  '/:bookId/reviews/:reviewId', 
+  authUser,
+  async (req, res, next) => {
+    res.setHeader('Content-Type', 'text/plain');
+    await axios.put(
+      'http://localhost:3001/reviews/7', 
+      {
+        username: req.user.name,
+        bookId: req.params.bookId,
+        review: req.body.review
+      }
+    );
+    res.status(200).send("Review Added.");
+  }
+);
 
 // Registered Users
 // Task 9: Delete book review added by that particular user
-booksRouter.delete('/reviews/:reviewId', async (req, res, next) => {
-  res.setHeader('Content-Type', 'text/plain');
-  await axios.delete(
-    `http://localhost:3001/reviews/${req.params.reviewId}`
-  );
-  res.status(200).send("Review Deleted.");
-});
+booksRouter.delete(
+  '/reviews/:reviewId', 
+  authUser,
+  async (req, res, next) => {
+    res.setHeader('Content-Type', 'text/plain');
+    const fullReview = await axios.get(
+      `http://localhost:3001/reviews/${req.params.reviewId}`
+    );
+    if(fullReview && fullReview.data.username != req.user.name) 
+      return res.status(403).send("Cannot delete another user's review.");
+    await axios.delete(
+      `http://localhost:3001/reviews/${req.params.reviewId}`
+    );
+    res.status(200).send("Review Deleted.");
+  }
+);
 
 module.exports = booksRouter;
